feat: Add social media icons
in this commit, i added fonts awesome social media icons to each team member. icons added are
facebook, twitter, github and briefcase icon for linking to each team member
porfolio site.
within the about-page.less file, the width of the container div was also set to 80% and
placed centrally on the screen