feat: Add products images and description:
In this commit, i added the products and services images and descritption. Two mobile screen images and a tablet image were
added to the products and services section. Three images were also added to the images folder. Images added are:
receipt_image1, receipt_image2 and receipt_image3.
Bug fix: The team image photo was also made to flow under the nav bar as against flowing on top of the nav bar