feat: Add Get started section
In this commit, i added a get started section to the about.html page. I also changed the
banner image to an image of a group of lambda students, (banner_image.jpg) which I belie
ve would be a better banner image for the receipt tracker team. css styles were also added
to the get started section and the call-to-action button within the get started section.