feat: Add more sections to about page
In this commit, a new section for additional team members was added to the about page.
also, a products description section and a meet-the-team header section was also added
to the about page. These were added to allow for some more description of the what the
receipt tracker app does and also to introduce more layout and structure to the entire
about web page in about.html

All newly added sections were also properly styled in the about-page.less