style: Merge header section in about.html
In this commit, i merged the header section of the about page with the header
section of Nathaniel Bosco's ( a team member ) header section. This is done
in order to achieve consistency across all web pages of the marketing website.

The header.less files were also updated to allow for custom css rules for mobile
and tablet screen widths
about-page.less was also refactored by nesting relevant css rule-sets together